refactor(#30): fix issues in naming and formatting
