+++ config/api_platform/resources.yaml
@@ -12,6 +12,6 @@ resources:
         openapiContext:
           responses:
             '204':
-               description: 'Health check passed with no content to return.'
+              description: 'Health check passed with no content to return.'
             '500':
-               description: 'Health check failed. One or more services are not functioning properly.'
+              description: 'Health check failed. One or more services are not functioning properly.'
